Added execute to default builds
